basic cellar dashboard
- create a new landing page with navigation
- move on tap to new /taps route
- create new /bottles endpoint
